refactor: move ExternalAuthProtocol to shared external_protocols module
