+++ backend/infrahub/auth/__init__.py
@@ -15,7 +15,8 @@ from infrahub.auth.auth import (
     validate_jwt_refresh_token,
 )
 from infrahub.auth.session import AccountSession
-from infrahub.auth.types import AuthType, ExternalAuthProtocol
+from infrahub.auth.types import AuthType
+from infrahub.external_protocols import ExternalAuthProtocol
 
 __all__ = [
     "AccountSession",
